+++ src/collection.assert.ts
@@ -7,11 +7,7 @@ import type {
   WithId,
 } from 'mongodb'
 import * as ta from 'type-assertions'
-import {
-  TsCollection,
-  TsReadCollection,
-  TsReadWriteCollection,
-} from './collection'
+import { TsCollection, TsReadCollection } from './collection'
 import type { TsFilter, TsFindOneAndDeleteOptions } from './types'
 
 type TSchema = { a: string; _id: ObjectId }
@@ -87,32 +83,3 @@ ta.assert<
     >
   >
 >()
-
-type DistinctSampleType = {
-  _id: ObjectId
-  str: string
-  nested: {
-    field: number
-  }
-}
-
-type TestCollection = TsReadWriteCollection<
-  DistinctSampleType & { writeOnly: boolean },
-  DistinctSampleType & { readOnly: boolean }
->
-
-const distinctString = (c: TestCollection) => c.distinct('str', {})
-const distinctId = (c: TestCollection) => c.distinct('_id', {})
-const distinctNested = (c: TestCollection) => c.distinct('nested.field', {})
-const distinctReadOnly = (c: TestCollection) => c.distinct('readOnly', {})
-
-ta.assert<ta.Equal<Awaited<ReturnType<typeof distinctString>>, string[]>>()
-ta.assert<ta.Equal<Awaited<ReturnType<typeof distinctId>>, ObjectId[]>>()
-ta.assert<ta.Equal<Awaited<ReturnType<typeof distinctNested>>, number[]>>()
-ta.assert<ta.Equal<Awaited<ReturnType<typeof distinctReadOnly>>, boolean[]>>()
-
-// @ts-expect-error
-const _distinctIncorrect = (c: TestCollection) => c.distinct('no', {})
-
-// @ts-expect-error
-const _distinctWriteOnly = (c: TestCollection) => c.distinct('writeOnly', {})
